+++ packages/signals_core/lib/src/core/signals.dart
@@ -2,6 +2,8 @@
 import 'dart:convert';
 import 'dart:developer' as developer;
 
+import 'package:signals_core/signals_core.dart';
+
 import 'observer.dart';
 import '../utils/constants.dart';